requirements.txt
agate==1.6.3
agate-dbf==0.2.2
agate-excel==0.2.5
agate-sql==0.5.8
amqp==5.1.1
appdirs==1.4.4
arabic-reshaper==2.1.4
asgiref==3.7.2
asn1crypto==1.5.1
attrs==22.1.0
autobahn==22.7.1
Automat==20.2.0
Babel==2.10.3
billiard==3.6.4.0
Brotli==1.0.9
cairocffi==1.4.0
CairoSVG==2.5.2
captcha==0.4
celery==5.2.7
certifi==2022.9.24
cffi==1.15.1
channels==3.0.5
charset-normalizer==2.1.1
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
constantly==15.1.0
cryptography==38.0.1
cssselect2==0.7.0
csvkit==1.0.7
daphne==3.0.2
dbfread==2.0.7
defusedxml==0.7.1
Django==4.2.7
django-admin-honeypot-advanced==1.0.1
django-admin-thumbnails==0.2.6
django-advanced-filters==2.0.0
django-celery-beat==2.3.0
django-celery-results==2.4.0
django-filter==22.1
django-ipware==4.0.2
django-recaptcha==3.0.0
django-timezone-field==5.0
django-weasyprint==2.2.0
djangorestframework==3.14.0
et-xmlfile==1.1.0
fonttools==4.37.3
fs==2.4.16
future==0.18.2
greenlet==1.1.3
html5lib==1.1
hyperlink==21.0.0
idna==3.4
incremental==21.3.0
isodate==0.6.1
Jinja2==3.1.2
kombu==5.2.4
leather==0.3.4
lxml==4.9.1
MarkupSafe==2.1.1
mysqlclient==2.2.0
olefile==0.46
openpyxl==3.0.10
oscrypto==1.3.0
parsedatetime==2.4
Pillow==9.2.0
prompt-toolkit==3.0.31
py-avataaars==1.1.2
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pydyf==0.3.0
pyHanko==0.14.0
pyhanko-certvalidator==0.19.5
pyOpenSSL==22.1.0
PyPDF3==1.0.6
pyphen==0.13.0
python-bidi==0.4.2
python-crontab==2.6.0
python-dateutil==2.8.2
python-decouple==3.6
python-slugify==6.1.2
pytimeparse==1.1.8
pytz==2022.2.1
pytz-deprecation-shim==0.1.0.post0
PyYAML==6.0
qrcode==7.3.1
razorpay==1.3.0
reportlab==3.6.11
requests==2.28.1
service-identity==21.1.0
short-url==1.2.2
simplejson==3.17.6
six==1.16.0
SQLAlchemy==1.4.41
sqlparse==0.4.4
svglib==1.4.1
text-unidecode==1.3
tinycss2==1.1.1
tqdm==4.64.1
Twisted==22.8.0
txaio==22.2.1
typing_extensions==4.8.0
tzdata==2022.2
tzlocal==4.2
uritools==4.0.0
urllib3==1.26.12
vine==5.0.0
wcwidth==0.2.5
weasyprint==56.1
webencodings==0.5.1
xhtml2pdf==0.2.8
xlrd==2.0.1
zope.interface==5.4.0
zopfli==0.2.1

